+++ R/plot.check_dag.R
@@ -143,7 +143,7 @@ plot.see_check_dag <- function(
 
   # these geoms are shared by both plots
   common_layers <- list(
-    geom_point_borderless(
+    geom_point_halo(
       ggplot2::aes(fill = .data$type),
       size = size_point
     ),